[op][agl][splash] Show immediately apps
First frame detection is not functional yet, even after porting the
Google Ozone Wayland backend fix.  As a workaround we reduce the time
waiting for the first frame to 300ms so it shows updates after first
300ms.

This is a temporary workaround. Later this should be reverted and a
proper fix should be added accordingly.

Bug-AGL: SPEC-4261