[agl][wayland] Update agl-shell protocol to version 2
Now clients need to wait for the bound_ok event to continue. From the
protocol documentation:

"Starting with version 2 of the protocol, the client is required to wait
for the 'bound_ok' or 'bound_fail' events in order to proceed further.

In case the client gets a 'bound_fail' event then it should consider that
there's another client already bound to the agl_shell protocol.
A client that receives a 'bound_ok' event should consider that there's
no other client already bound to the interface and can proceed
further."